// src/models/responses/DashboardModels.ts
export interface IDashboard {
  totalRegistrations: number;
  totalBeds: number;
  totalUsedBeds: number;
  totalEmptyBeds: number;
  totalUsers: number;
  totalCurrentUsers: number;
  totalRequests: number;
  totalParkingRequests: number;
  totalUnResovledRequests: number;
  totalUnResovledParkingRequests: number;
  totalMaleUsers: number;
  totalFemaleUsers: number;
  contractsPerMonths: number[];
}